Rc (#26)
* new helper scdriupt added

* Add files via upload

* Delete skripts/export/common.conf

* Delete skripts/export/common_funcs.sh

* Delete skripts/export/loits2.sh

* few more pkgs added

* cmdline parsing maybe works

* ... works with loits_new i mean

* enforce_deps kinda works

* inst_dep kinda wks, s0rt 0f

* Add files via upload

* sudo-related

* still trying 2 ake loits_new work

* grub-mkrescue

* Add files via upload

* fixing sme things

* trying 2 make loits2 w0rk

* loits2 opt10n pars1ng

* l2.check.pars.1

* ls2cp3

* sco,scm

* debug added 2 loits2

* make_tar()

* mostly abt mkconffile.sh

* Delete skripts/export/mkconffile.sh.export

* Add files via upload

* new file

* Delete stage0.sh.export

* new file

* Delete skripts/export/stage0_backend.bsh.export

* try1ng 2 mak3 backend w0rk again

* mostly abt commments

* comments2

* mostly abt make_tgt_dirs()

* mostly abt make_src_dirs()

* trying 2 guess what i meant by "make_dirs"

* mostly abt "sudo cp"

* mostöyu abt "sudo rm"

* Add files via upload

* Delete stage0f.sh.export

* debug

* mostly abt sudo

* new stuff

* new stuff 2

* Add files via upload

* Add files via upload

* Add files via upload

* Add files via upload

* i hope these scripts are in "almost works" state so that

... i can soon test with an actual .iso-file

* i hope...

* vommentz

* Update common_funcs.sh

vommentz abt future devel

* trying 2 make stage0+stage0f work

* trying 2 make stage0+stage0f work

* Add files via upload

* Add files via upload

* maybe bootloader()

* maybe bootloader() works

... sho that i can concentrate on "loits"-scripts

* changing how to initialize tha "dev_env"

* conf_gi_opts changed for loits_skripts

* changes

* changes.2

i hope that loits_new starts to work soon

* init-dev-env-related

* saturday's changes

* saturday's changes 2

- i hope that loits_new's output boots now
(not yet 100% because Reasons)

- also some changes to "init-scripts"

* tried 2 mak3 loits2 work

Not exactly sure why there are 2 different scripts for same job. 
Was there an actual reason 4 that or not? Maybe i was just bored and decided to write a new version.

* dependencies

* sunday's changes 2

* sunday's changes 3

* commentsa

* cleanup

* cleanup

* maybe

* deps

actually i'm supposed to make new .iso-files boot, instead stick on doing sometrging else

* deps again

* grub almost works

actually  tonighti made a grub-bootable .iso once, by accident
next;make grub-bootable happen on purpose

* grub almost works+other changes

* pkg mtools retunred

* monday's changes 1

* monday's changes 2

* Delete skripts/export/mksums_new.sh.export

* Delete skripts/export/common_funcs.sh.export

* Delete skripts/export/loits2.sh.export

* Delete skripts/export/loits_new.sh

* Delete skripts/export/loits_new.sh.export

* Add files via upload

* maybe mksums kinda?

* friday's changes 1

* pkg install changed a bit

* new script added

may not tey work properly

* Delete skripts/export/squ.ash.export

* mostly abt mksums

* freya's dag 2

* changes here and there

maybe mksums works properly now?
removed debug
changed grub-spesific stuff

* lördagur

* mostly abt squ.ash

maybe some options work now

* abt squ.ash

more options

* squ.ash + let's find out

* maybe -f, -b

* baron Monday

* baron Monday.2

* tuesday.1

* tuesday.2

* tuesday.2

* tuesday.3

+++ skripts/squ.ash
@@ -10,9 +10,11 @@ cmd=""
 md=0
 mp=0
 ms=0
+par=""
 
 function parse_opts_real() {
-	echo "squash.parse_opts_real()"
+	echo "squash.parse_opts_real()" #dqb
+
 
 	case ${1} in
 		--dir2)
@@ -51,8 +53,14 @@ parse_opts ${7} ${8}
 parse_opts ${9} ${10}
 
 function xxx() {
+
+	dqb "xxx( ${1})"
+	#tulisi stopata tässä jos ei kalaa
+	[ -s ${1} ] || exit 99
+
 	if [ x"${CONF_squash0}" != "x" ]; then
-		[ -d ${CONF_squash0} ] || ${smd}${CONF_squash0}
+		[ -d ${CONF_squash0} ] || ${smd} ${CONF_squash0}
+
 		cd ${CONF_squash0}
 		unsq=$(${odio} which unsquashfs)
 
@@ -63,81 +71,91 @@ function xxx() {
 		fi
 
 	fi
+
+
+	dqb "xxx d0mw"
 }
 
-function fix_sudo() {
-	if [ x"${CONF_squash_dir}" != "x" ]; then
-		cd ${CONF_squash_dir}
-
-		${sco} -R 0:0 ./etc/sudo*
-		${scm} -R a-w ./etc/sudo*
-		${sco} -R 0:0 ./usr/lib/sudo/*
-
-		#${sco} -R 0:0 ./usr/bin/sudo*
-		#RUNNING SOME OF THESE COMMANDS OUTSIDE CHROOT ENV STARTED TO SEEM LIKE A BAD IDEA
-		#AND CHATTR MAY OT WORK WITH SOME FILESYSTEMS	
-
-		${scm} 0750 ./etc/sudoers.d
-		${scm} 0440 /etc/sudoers.d/*
-
-		${scm} -R a-w ./usr/lib/sudo/*
-		#${scm} -R a-w ./usr/bin/sudo*
-		#${scm} 4555 ./usr/bin/sudo
-		${scm} 0444 ./usr/lib/sudo/sudoers.so
-
-		#${sca} +ui ./usr/bin/sudo
-		#${sca} +ui ./usr/lib/sudo/sudoers.so	
-	fi
-}
-
-function bbb() {
-
-	if [ x"${CONF_squash_dir}" != "x" ]; then
-
-		if [ -d ${CONF_squash_dir} ] ; then 
-			cd ${CONF_squash_dir}
-
-			${smr} -rf ./run/live
-			${smr} -rf ./boot/grub/*
-			${smr} -rf ./boot/*
-			${smr} -rf ./usr/share/doc/*	
-			${smr} -rf ./var/cache/apt/archives/*.deb
-			${smr} -rf ./var/cache/apt/*.bin
-			${smr} -rf ./tmp/*
-
-			fix_sudo
-
-			${scm} -R 0755 ./var/cache/man
-			${sco} -R man:man ./var/cache/man
-
-			${smr} ./root/.bash_history
-			${smr} ./home/devuan/.bash_history
-			
-		fi
-	fi
-
-}
+#jatkossa common_lib?
+#tilapäisesti jemmmaan koska pykimistä slim:in kanssa TAAS
+#function fix_sudo() {
+#	if [ x"${CONF_squash_dir}" != "x" ]; then
+#		cd ${CONF_squash_dir}
+#		[ ${debug} -eq 1 ] && pwd
+#		csleep 1 
+#
+#		${sco} -R 0:0 ./etc/sudo*
+#		${scm} -R a-w ./etc/sudo*
+#		${sco} -R 0:0 ./usr/lib/sudo/*
+#
+#		#${sco} -R 0:0 ./usr/bin/sudo*
+#		#RUNNING SOME OF THESE COMMANDS OUTSIDE CHROOT ENV STARTED TO SEEM LIKE A BAD IDEA
+#		#AND CHATTR MAY OT WORK WITH SOME FILESYSTEMS	
+#
+#		${scm} 0750 ./etc/sudoers.d
+#		${scm} 0440 /etc/sudoers.d/*
+#
+#		${scm} -R a-w ./usr/lib/sudo/*
+#		#${scm} -R a-w ./usr/bin/sudo*
+#		#${scm} 4555 ./usr/bin/sudo
+#		${scm} 0444 ./usr/lib/sudo/sudoers.so
+#
+#		#${sca} +ui ./usr/bin/sudo
+#		#${sca} +ui ./usr/lib/sudo/sudoers.so	
+#	fi
+#}
+#
+#function bbb() {
+#	dqb "bbb()"
+#
+#	if [ x"${CONF_squash_dir}" != "x" ]; then
+#
+#		if [ -d ${CONF_squash_dir} ] ; then 
+#			#exit 55
+#			cd ${CONF_squash_dir}
+#			[ ${debug} -eq 1 ] && pwd
+#
+#			${smr} -rf ./run/live
+#			${smr} -rf ./boot/grub/*
+#			${smr} -rf ./boot/*
+#			${smr} -rf ./usr/share/doc/*	
+#			${smr} -rf ./var/cache/apt/archives/*.deb
+#			${smr} -rf ./var/cache/apt/*.bin
+#			${smr} -rf ./tmp/*
+#
+#			#fix_sudo #tälle tarttis tehdä jotain ettei ala pykumään kun luotu kiekko bootattu
+#
+#			${scm} -R 0755 ./var/cache/man
+#			${sco} -R man:man ./var/cache/man
+#
+#			${smr} ./root/.bash_history
+#			${smr} ./home/devuan/.bash_history	
+#		fi
+#	fi
+#
+#	dqb "bbb().done()"
+#}
 
 function jlk_main() {
+	dqb "jkl1 $1 "
+	[ x"${1}" != "x" ] || exit 66
 
 	if [ x"${CONF_squash_dir}" != "x" ]; then
 		echo "${0} -x ?"
-		[ -d ${CONF_squash_dir}/${TARGET_pad2} ] || ${smd}-p ${CONF_squash_dir}/${TARGET_pad2}
+		[ -d ${CONF_squash_dir}/${TARGET_pad2} ] || ${smd} -p ${CONF_squash_dir}/${TARGET_pad2}
 
 		cd ${CONF_squash_dir}/${TARGET_pad2}
-		[ x"${1}" != "x" ] || exit 666
-
+		[ ${debug} -eq 1 ] && pwd
+		csleep 1
 
 		local d
 		for d in sh bz2 bz3 ; do ${spc} ${1}/${TARGET_pad_dir}/*.${d} . ; done
-
-
-
 	
 	fi
 }
 
 function jlk_conf() {
+	dqb "jlk_conf( ${1} , ${2})"
 
 	if [ x"${CONF_squash_dir}" != "x" ] && [ y"${1}" != "y" ] ; then
 		if [ -d ${1}/${TARGET_pad_dir} ] ; then 
@@ -167,12 +185,14 @@ function jlk_conf() {
 
 function jlk_sums() {
 
+	dqb "jlk_sums( ${1} , ${2})"
+
 	[ x"${1}" != "x" ] || exit 666
 	[ -d ${1} ] || echo "no such thing as ${1}"
 
 	cd ${CONF_squash_dir}/${TARGET_pad2}
+	[ -d ./0 ] || ${smd} -p ./0 ;sleep 6
 
-	[ -d ./0 ] || ${smd}-p ./0 ;sleep 6
 
 
 	${spc} -a ${1}/${TARGET_DIGESTS_dir}/* ./0
@@ -185,30 +205,51 @@ function jlk_sums() {
 
 function rst() {
 
-	if [ x"${CONF_squash_dir}" != "x" ]; then
-		cd ${CONF_squash_dir}
+	dqb "rst( ${1} , ${2} )"
 
-		if [ ${md} -gt 0 ] ; then
-			${som} -o bind /dev ./dev 
+
+	if [ x"${CONF_squash_dir}" != "x" ]; then
+		    cd ${CONF_squash_dir}
+
+
+			  dqb "MOUNTING PTOC"
+
+			  ${som} -o bind /dev ./dev 
 		fi
 
 		if [ ${ms} -gt 0 ] ; then
+
+			dqb "MOUNTING ssy"
+
 			${som} -o bind /sys ./sys
 		fi
 
 		if [ ${mp} -gt 0 ] ; then
+			dqb "MOUNTING PC9EOR"
+
 			${som} -o bind /proc ./proc 
 			${odio} ln -s ./proc/mounts ./etc/mtab
 		fi		 
 
-		${svm} ${CONF_squash_dir}/etc/hosts ${CONF_squash_dir}/etc/hosts.bak	
+		pwd
+		csleep 5
+
+		#TODO:jotain säätöä tämän ympäiorstln kanssa ok buelä
+		#... esim /e/d/locale voisio lla hyväkopsata perlin valitusten johdosra
+
+		[ -f ${CONF_squash_dir}/etc/hosts ]  && ${svm} ${CONF_squash_dir}/etc/hosts ${CONF_squash_dir}/etc/hosts.bak	
 		${spc} /etc/hosts ${CONF_squash_dir}/etc
-		
+		${odio} touch ./.chroot
+		#date > ./.chroot
+
 		${odio} chroot ./ ./bin/bash 
-				
+		[ $? -eq 0 ] || echo "MOUNT -O REMOUNT,EXEC ${CONF_tmpdir0}"
+		
+		${smr} ./.chroot			
 		${svm} ./etc/hosts.bak ./etc/hosts
 		${smr} ./etc/mtab
-		[ $? -eq 0 ] || echo "MOUNT -O REMOUNT,EXEC ${CONF_tmpdir0}"
+
+
 		sleep 3
 	
 		${uom} ./dev 
@@ -217,16 +258,23 @@ function rst() {
 
 		sleep 3
 	fi
-
 }
 
 function cfd() {
+	dqb "cfd(${1})"
+	[ x"${1}" != "x" ] || exit 6
+	[ -s ${1} ] && exit 66
+	dqb "PARS IJ"
+
+
 	if [ x"${CONF_squash_dir}" != "x" ]; then
 		echo "${0} -b ?"
 
 		cd ${CONF_squash_dir}
 
 		[ -s ${1} ] && rm ${1}
+		local msq
+
 		msq=$(${odio} which mksquashfs)
 
 		if [ x"${msq}" != "x" ] && [ -x ${msq} ] ; then 
@@ -235,53 +283,56 @@ function cfd() {
 			echo "${odio} apt-get install squashfs-utils"
 		fi
 	fi
+
+	dqb "cfd(${1}) DONE"
 }
 
 function usage() {
-	echo "-x <source_file>:eXtracts <source_file> to ${CONF_squash0}"
-	echo "-y <iso_file> extracts <iso_file>/live/filesystem.squashfs to ${CONF_squash0}"
-	echo "-b is supposed to be run just Before -c but after -r"
+	echo "-x <source_file>:eXtracts <source_file> to ${CONF_squash0} source_file NEEDS TO HAVE ABSOLUTE PATH"
+	echo "-y <iso_file> extracts <iso_file>/live/filesystem.squashfs to ${CONF_squash0} NEEDS TO HAVE ABSOLUTE PATH"
+	echo "-b is supposed to be run just Before -c but after -r (TODO)"
 	echo "-d Destroys contents of ${CONF_squash0}/ and ${CONF_tmpdir} "
 	echo "\t if possible, use -b instead\n"
 
-	echo "-c <target_file> [optional_params_4_mksquashfs?] : Compresses ${CONF_squash_dir} to <target_file> (with optional_params?)\n"
-	echo "-r :runs chRoot in ${CONF_squash_dir}"
+	echo "-c <target_file> [optional_params_4_mksquashfs?] : Compresses ${CONF_squash_dir} to <target_file> (with optional_params?)\n NEEDS TO HAVE ABSOLUTE PATH"
+	echo "-r :runs chRoot in ${CONF_squash_dir} "
 	
-	echo "-i <src> : Installs scripts 2 chroot_dir "
-	echo "-j <src> [ --dir2 <stuff> ] : extracts dir 2 chroot_dir"
+	#echo "-i <src> : Installs scripts 2 chroot_dir  (TODO?)"
+	echo "-j <src> [ --dir2 <stuff> ] : extracts dir 2 chroot_dir  NEEDS TO HAVE ABSOLUTE PATH"
+
 	echo "\t to state the obvious:"
 	echo "\t <stuff> in --dir2 has to contain sub-directory ${TARGET_DIGESTS_dir} , for example ${CONF_target} "
 	echo "\t ... and <src> has to contain subdirectory ${TARGET_pad_dir} \n"
 
-	echo "-f attempts to Fix some problems w/ sudo"
+	echo "-f attempts to Fix some problems w/ sudo "
 	echo "--v 1 adds Verbosity\n"
 
-	echo "--mp,--md, --ms:self-explanatory opts 4 -r" 
+	echo "--mp,--md, --ms:self-explanatory opts 4 -r (TODO)" 
 	echo "\t potentially dangerous, so disabled by default , 1 enables"
 }
 
-function ijk() {
-	if [ x"${CONF_squash_dir}" != "x" ]; then
-		echo "${0} -x ?"
-		
-		previous=$(pwd)
-		${smd}-p ${CONF_squash_dir}/${TARGET_pad2}
-		cd ${CONF_squash_dir}/${TARGET_pad2}
-	
-		[ x"${1}" != "x" ] || exit 666
-		[ -d ${1} ] || exit 666
-
-		local d
-		for d in ${TARGET_patch_list_2} ; do ${spc} ${1}/${TARGET_pad_dir}/${d} . ; done
-
-		grep -v TARGET_to_ram devuan.conf > root.conf 
-	
-		${smr} devuan.conf
-		echo "TARGET_to_ram=1" >> root.conf
-		${scm} 0444 root.conf
-		cd ${previous}
-	fi
-}
+#function ijk() {
+#	if [ x"${CONF_squash_dir}" != "x" ]; then
+#		echo "${0} -x ?"
+#		
+#		previous=$(pwd)
+#		${smd} -p ${CONF_squash_dir}/${TARGET_pad2}
+#		cd ${CONF_squash_dir}/${TARGET_pad2}
+#	
+#		[ x"${1}" != "x" ] || exit 666
+#		[ -d ${1} ] || exit 666
+#
+#		local d
+#		for d in ${TARGET_patch_list_2} ; do ${spc} ${1}/${TARGET_pad_dir}/${d} . ; done
+#
+#		grep -v TARGET_to_ram devuan.conf > root.conf 
+#	
+#		${smr} devuan.conf
+#		echo "TARGET_to_ram=1" >> root.conf
+#		${scm} 0444 root.conf
+#		cd ${previous}
+#	fi
+#}
 
 case ${cmd} in
 	-x)
@@ -289,13 +340,24 @@ case ${cmd} in
 	;;
 	-y)
 		[ -d ${CONF_source} ] || ${smd} -p ${CONF_source}
-		${som} -o loop,ro ${par} ${CONF_source}
-		xxx ${CONF_source}/live/filesystem.squashfs
-		${uom} ${CONF_source}
+		dqb "${som} -o loop,ro ${par} ${CONF_source}"
+
+		oldd=$(pwd)
+		${som} -o loop,ro ${par} ${oldd}/${CONF_source}
+		[ ${debug} -eq 1 ] && ls -las ${oldd}/${CONF_source}/live/
+		csleep 3
+
+		#[ ${debug} -eq 1 ] && dirname $0
+		[ ${debug} -eq 1 ] && pwd
+		csleep 3
+
+		xxx ${oldd}/${CONF_source}/live/filesystem.squashfs
+		${uom} ${oldd}/${CONF_source}
 	;;
-	-b) 
-		bbb		
-	;;
+#	-b) 
+#		bbb		
+#	;;
+
 	-d)
 
 		if [ x"${CONF_squash0}" != "x" ] ; then
@@ -315,6 +377,9 @@ case ${cmd} in
 		rst
 	;;
 	-j)
+
+		#sudo: unable to allocate pty: No such device
+
 		jlk_main ${par}
 
 		[ z"${dir2}" != "z" ] || echo "--dir2 "
@@ -322,11 +387,13 @@ case ${cmd} in
 		jlk_conf ${dir2} devuan
 
 		jlk_sums ${dir2}
-		fix_sudo
+
+		#fix_sudo
 	;;
-	-f)
-		fix_sudo
-	;;
+#	-f)
+#		fix_sudo
+#	;;
+
 	*)
 		usage
 	;;
